Fix androidx.startup related code
Better structure all initializers so that they properly depend on each
other while also allowing DatadogInitializer.kt to be in its own module,
 by making it abstract and resolving its dependencies in the :app module
