+++ app/src/main/AndroidManifest.xml
@@ -322,7 +322,10 @@
                 android:name="com.hedvig.app.feature.tracking.LaunchTrackingInitializer"
                 android:value="androidx.startup" />
             <meta-data
-                android:name="com.hedvig.android.datadog.DatadogInitializer"
+                android:name="com.hedvig.app.startup.DatadogInitializerImpl"
+                android:value="androidx.startup" />
+            <meta-data
+                android:name="com.hedvig.app.startup.TimberInitializer"
                 android:value="androidx.startup" />
         </provider>
     </application>
