+++ app/src/main/kotlin/com/hedvig/app/startup/TimberInitializer.kt
@@ -1,0 +1,22 @@
+package com.hedvig.app.startup
+
+import android.content.Context
+import androidx.startup.Initializer
+import com.hedvig.app.isDebug
+import com.hedvig.app.util.firebase.FirebaseBreadcrumbTimberTree
+import com.hedvig.app.util.firebase.FirebaseCrashlyticsLogExceptionTree
+import timber.log.Timber
+
+class TimberInitializer : Initializer<Unit> {
+  override fun create(context: Context) {
+    if (isDebug()) {
+      Timber.plant(Timber.DebugTree())
+    }
+    Timber.plant(FirebaseBreadcrumbTimberTree())
+    Timber.plant(FirebaseCrashlyticsLogExceptionTree())
+  }
+
+  override fun dependencies(): List<Class<out Initializer<*>>> {
+    return emptyList()
+  }
+}
